fix: 修复类型推断与 int48 序列化/范围截断问题
- 语义: new Struct(named=params) 命名参数字段访问(p1.x)被推断为 any。visit_struct_init 先推断后 visit，AST_MODULE_ACCESS 未转换为 AST_FIELD_ACCESS，infer 落入 default 返回 TYPE_ANY。改为先 visit 后推断。

- 序列化: CONST_TAG_INT 用 int32 存储导致 >= 2^31 的 int48 值被截断。f9ec68f4 让 .leno 直走序列化后激活此潜在 bug；改用 u64 读写，递归覆盖 struct 字段默认值/数组/字典。

- 范围: ObjRange 全链路 int32 截断(VM/struct/序列化)，与 int48 模型不一致。start/end 改 int64_t，OP_RANGE 用 val_as_int 精确取值，bigint 越界抛清晰错误；CONST_TAG_RANGE 同步改 u64。

- LENO_BIN_VERSION 升至 v2.2.0 作废旧 .lenb 缓存。

+++ src/include/leno_value.h
@@ -428,8 +428,8 @@ typedef struct {
 // 范围对象
 typedef struct {
     Object header;
-    int start;          // 起始值
-    int end;            // 结束值
+    int64_t start;      // 起始值（int48，用 int64_t 容器存储，与 Leno int 模型对齐）
+    int64_t end;        // 结束值（int48，用 int64_t 容器存储，与 Leno int 模型对齐）
     int inclusive;      // 是否包含结束值（1=包含，0=不包含）
 } ObjRange;
 
@@ -844,8 +844,8 @@ static inline ObjString* str_copy(const char* chars, int len) {
 ObjString* str_concat(ObjString* a, ObjString* b);
 
 // Range 操作
-ObjRange* range_new(int start, int end, int inclusive);
-int range_contains(ObjRange* range, int value);
+ObjRange* range_new(int64_t start, int64_t end, int inclusive);
+int range_contains(ObjRange* range, int64_t value);
 
 // BigInt 操作
 ObjBigInt* bigint_new(const uint32_t* limbs, int limb_count, int is_negative);
